required optimized  indexing fix
